Add attach aws session stack command
This command adds the ability to attach write-only and/or read-only credentials temporarily to a stack from your local environment. This can be helpful in situations where the spacelift integration does not have full administrative credentials for some resources for security reasons, and you need to temporarily grant higher privileges to an individual spacelift run.